backend/requirements.txt
# Backend Dependencies for Smart DevCopilot

# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# AI/ML Libraries
openai==1.3.0
langchain==0.0.335
transformers==4.35.0
torch>=2.6.0
sentence-transformers==2.2.2

# Vector Databases
faiss-cpu==1.7.4
# pinecone-client==2.2.4  # Uncomment if using Pinecone

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.1
aiofiles==23.2.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and Logging
prometheus-client==0.19.0
python-logging-loki==0.3.1
